app/requirements.txt
langchain
tiktoken
openai
langsmith
python-dotenv
weaviate-client
langserve
sse_starlette
pydantic
python-telegram-bot

